Improve Windows NSIS installer script (setup.nsi)
Two improvements of installer script :
- avoid error message when Unbound is running,
- add "DisplayVersion" in registry thus Windows package manager (Winget) can handle Unbound.

Avoid error message :
When Unbound is already installed and running, when the installer is used to update Unbund, it displays an error message  "Error opening file for writing".
In order to avoid this message, I suggest a modification to stop Unbound service if already installed.

Adding "DisplayVersion" in registry :
Winget use uninstall parameters in Windows registry to get program versions to be able to compare installed version and last available version to propose update. Without this parameter, Winget can't handle Unbound updates. This parameter is also used to show program versions in the Windows remove program menu.
I suggest adding this parameter in the installer script.